==> 3_data_enrichment.py <==
# -*- coding: utf-8 -*-
"""3_data_enrichment.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1qAANvqhRJAhaJa7tF40Ex3b4utHL08nk
"""

# Import standard libraries
import os
import pathlib

# Import installed libraries
import pandas as pd

try:
  from google.colab import drive
  drive.mount("/gdrive")
  core_path = "/gdrive/MyDrive/L&D/Own projects/Space science with Python/AsteroidSpectra/asteroid_taxonomy"
except ModuleNotFoundError:
  core_path = ""

# Read the level 1 dataframe
asteroids_df = pd.read_pickle(os.path.join(core_path, "data/lvl1/", "asteroids_merged.pkl"))

# Create a dictionary that maps the Bus Classification with the main group
bus_to_main_dict = {
                    'A': 'Other',
                    'B': 'C',
                    'C': 'C',
                    'Cb': 'C',
                    'Cg': 'C',
                    'Cgh': 'C',
                    'Ch': 'C',
                    'D': 'Other',
                    'K': 'Other',
                    'L': 'Other',
                    'Ld': 'Other',
                    'O': 'Other',
                    'R': 'Other',
                    'S': 'S',
                    'Sa': 'S',
                    'Sk': 'S',
                    'Sl': 'S',
                    'Sq': 'S',
                    'Sr': 'S',
                    'T': 'Other',
                    'V': 'Other',
                    'X': 'X',
                    'Xc': 'X',
                    'Xe': 'X',
                    'Xk': 'X'
                   }

# Create a new "main group class"
asteroids_df.loc[:, "Main_Group"] = asteroids_df["Bus_Class"].apply(lambda x:
                                                                    bus_to_main_dict.get(x, "None"))

# Remove the file path and Designation Number
asteroids_df.drop(columns=["DesNr", "FilePath"], inplace=True)

# Show the final data set for anyone who is interested ...
asteroids_df

# ... and also the spectrum of Ceres
asteroids_df.loc[asteroids_df["Name"] == "1 Ceres"]["SpectrumDF"][0]

# Create Level 2 directory and save the dataframe
pathlib.Path(os.path.join(core_path, "data/lvl2")).mkdir(parents=True, exist_ok=True)

# Save the dataframe as a pickle file
asteroids_df.to_pickle(os.path.join(core_path, "data/lvl2/", "asteroids.pkl"), protocol=4)
